break strmap into its own module
